Use toml-test parser interface, move some tests to there
Expand on #297 by not calling the toml-test-* binaries with the Parser
interface added in https://github.com/BurntSushi/toml-test/pull/73

Also move a bunch of tests to toml-test.
