+++ .github/workflows/test.yml
@@ -1,22 +1,30 @@
-on: ['push', 'pull_request']
-name: 'go test'
-
-jobs:
-  test:
-    strategy:
-      matrix:
-        go-version: ['1.13.x', '1.14.x', '1.15.x', '1.16.x']
-        os:         ['ubuntu-latest', 'macos-latest', 'windows-latest']
-    runs-on: '${{ matrix.os }}'
-    steps:
-
-    - name: 'Install Go'
-      uses: 'actions/setup-go@v2'
-      with:
-        go-version: '${{ matrix.go-version }}'
-
-    - name: 'Checkout code'
-      uses: 'actions/checkout@v2'
-
-    - name: 'Test'
-      run: 'make'
+{
+  "name": "go test",
+  "on":   ["push", "pull_request"],
+  "jobs": {
+    "test": {
+      "strategy": {
+        "matrix": {
+          "go-version": ["1.13.x", "1.14.x", "1.15.x", "1.16.x"],
+          "os":         ["ubuntu-latest", "macos-latest", "windows-latest"]
+        }
+      },
+      "runs-on": "${{ matrix.os }}",
+      "steps": [
+        {
+          "name": "Install Go",
+          "uses": "actions/setup-go@v2",
+          "with": {"go-version": "${{ matrix.go-version }}"}
+        },
+        {
+          "name": "Checkout code",
+          "uses": "actions/checkout@v2"
+        },
+        {
+          "name": "Test",
+          "run": "go test -v ./..."
+        }
+      ]
+    }
+  }
+}
