+++ go.mod
@@ -2,4 +2,4 @@ module github.com/BurntSushi/toml
 
 go 1.13
 
-require github.com/BurntSushi/toml-test v0.1.1-0.20210619051135-67b199419f32
+require github.com/BurntSushi/toml-test v0.1.1-0.20210620192437-de01089bbf76
